convert few parts to TS
